+++ src/main/php/peer/ftp/FtpEntryList.class.php
@@ -1,13 +1,15 @@
 <?php namespace peer\ftp;
 
+use IteratorAggregate, Traversable;
 use lang\Value;
+use util\Objects;
 
 /**
  * List of entries on an FTP server
  *
  * @see   xp://peer.ftp.FtpDir#entries
  */
-class FtpEntryList implements Value, \IteratorAggregate {
+class FtpEntryList implements Value, IteratorAggregate {
   protected
     $connection   = null,
     $list         = [],
@@ -26,13 +28,8 @@ class FtpEntryList implements Value, \IteratorAggregate {
     $this->base= $base;
   }
   
-  /**
-   * Returns an iterator for use in foreach()
-   *
-   * @see     php://language.oop5.iterations
-   * @return  php.Iterator
-   */
-  public function getIterator() {
+  /** Iterators over all entries */
+  public function getIterator(): Traversable {
     return new FtpListIterator($this->list, $this->connection, $this->base);
   }
 
@@ -75,7 +72,7 @@ class FtpEntryList implements Value, \IteratorAggregate {
 
   /** @return string */
   public function toString() {
-    return nameof($this).'('.$this->size().' entries)@'.\xp::stringOf($this->list);
+    return nameof($this).'('.$this->size().' entries)@'.Objects::stringOf($this->list);
   }
 
   /**