+++ ChangeLog.md
@@ -3,6 +3,12 @@ FTP protocol support for the XP Framework ChangeLog
 
 ## ?.?.? / ????-??-??
 
+## 10.0.1 / 2021-09-24
+
+* Fixed PHP 8.0 and PHP 8.1 compatibility - @thekid
+* Fixed connection to FTP servers with multiline banner message - @thekid
+* Replaced deprecated xp::stringOf() calls with util.Objects - @thekid
+
 ## 10.0.0 / 2020-04-10
 
 * Implemented xp-framework/rfc#334: Drop PHP 5.6: